Refactoring calls to change Category Id's to names
