+++ ProcessMaker/Models/ProcessCategory.php
@@ -68,4 +68,19 @@ class ProcessCategory extends ProcessMakerModel
     {
         return $this->morphedByMany(Process::class, 'assignable', 'category_assignments', 'category_id');
     }
+
+    /**
+     * Get Process Category Names
+     * @param string String of ids separated by a custom delimiter.
+     * @param string Delimiter to split ids. By default ','
+     * @return string A string separated by commas with Process Category Names
+     */
+    public static function getNamesByIds(string $ids, string $delimiter = ','): string
+    {
+        $resultString = '';
+        $arrayIds = explode($delimiter, $ids);
+        $results = ProcessCategory::whereIn('id', array_map('intval', $arrayIds))->pluck('name');
+        $resultString = implode(', ', $results->toArray());
+        return $resultString;
+    }
 }
