+++ tests/unit/ProcessMaker/Helpers/ArrayHelperTest.php
@@ -71,62 +71,4 @@ class ArrayHelperTest extends TestCase
             ]
         ));
     }
-
-    public function testGetNamesByIds()
-    {
-        //Case 1: Existing Model, Existing Name, More than one ID
-        $stringModel = 'ProcessCategory';
-        $stringIds = '1,2';
-        $columnName = 'name';
-        $this->assertEquals(
-            "Default Templates, System",
-            ArrayHelper::getNamesByIds($stringModel, $stringIds, $columnName)
-        );
-
-        //Case 2: Existing Model, Existing Name, one ID
-        $stringModel = 'ProcessCategory';
-        $stringIds = '1';
-        $columnName = 'name';
-        $this->assertEquals(
-            "Default Templates",
-            ArrayHelper::getNamesByIds($stringModel, $stringIds, $columnName)
-        );
-
-        //Case 3: Not Existing Model, Existing Name, More than one ID
-        $stringModel = 'ProcessCategoryFake';
-        $stringIds = '1,2';
-        $columnName = 'name';
-        $this->assertEquals(
-            "",
-            ArrayHelper::getNamesByIds($stringModel, $stringIds, $columnName)
-        );
-
-        //Case 4: Not Existing Model, Not Existing Name, More than one ID
-        $stringModel = 'ProcessCategoryFake';
-        $stringIds = '1,3';
-        $columnName = 'nameFake';
-        $this->assertEquals(
-            "",
-            ArrayHelper::getNamesByIds($stringModel, $stringIds, $columnName)
-        );
-
-        //Case 5: Not Existing Model, Not Existing Name, Not existing ID
-        $stringModel = 'ProcessCategoryFake';
-        $stringIds = '';
-        $columnName = 'nameFake';
-        $this->assertEquals(
-            "",
-            ArrayHelper::getNamesByIds($stringModel, $stringIds, $columnName)
-        );
-
-        //Case 6: Not Existing Package
-        $stringModel = 'ProcessCategoryFake';
-        $stringIds = '';
-        $columnName = 'nameFake';
-        $packageName = 'packageFake';
-        $this->assertEquals(
-            "",
-            ArrayHelper::getNamesByIds($stringModel, $stringIds, $columnName, $packageName)
-        );
-    }
 }
